Take over class loading capabilities.
Other classloaders only interfere, too many cooks, this causes the class already loaded or duplicate class issues.
The perfectLoaders gets created on construction, a collection of , and is used to optimize the verification of the current spl_autoloaders. This check is only done on create and if a class was not found as per normal, iow we are considering further investigation.
The autoloaders are unregistered but kept in collection and should it happen that we are unsuccessful yet we can as last resort call the autoloaders to see if the class may be resolved after all.